Mern stack complete rest api functionality

+++ backend/controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc Get Goals
 // @route Get /api/goals
@@ -18,6 +19,7 @@ const setGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -31,6 +33,18 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  const user = await User.findById(req.user.id);
+  // Check for the right user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  // Compare the user id to user goal
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndRemove(req.params.id, req.body, {
     new: true,
   });
@@ -40,6 +54,23 @@ const updateGoal = asyncHandler(async (req, res) => {
 // @route Delete /api/goals
 // @access Private
 const deleteGoal = asyncHandler(async (req, res) => {
+  const goal = await Goal.findById(req.params.id);
+  if (!goal) {
+    res.status(400);
+    throw new Error("Goal not found");
+  }
+  // Find the right user
+  const user = await User.findById(req.user.id);
+  // Check for the right user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  // Compare the user id to user goal
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
   await Goal.findByIdAndDelete(req.params.id);
   res.status(200).json(req.params);
 });
